pillmate/mobile_backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# AWS / DynamoDB
boto3>=1.34.0

# Auth
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# Google Gemini
google-generativeai>=0.7.0

# HTTP client (for FDA API)
httpx>=0.27.0

# Config
python-dotenv>=1.0.0

# Pydantic (FastAPI depends on v2)
pydantic[email]>=2.7.0
